+++ dev-java/openjdk-bin/openjdk-bin-8.312_p07.ebuild
@@ -1,11 +1,11 @@
 # Copyright 1999-2021 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
-EAPI=6
+EAPI=8
 
-inherit eapi7-ver java-vm-2
+inherit java-vm-2
 
-MY_ZULU_PV="8.56.0.21-ca-jdk8.0.302"
+MY_ZULU_PV="8.58.0.13-ca-jdk8.0.312"
 MY_ZULU_ARCH="linux_musl_x64"
 MY_PV=${PV/_p/+}
 SLOT=${MY_PV%%[.+]*}
@@ -19,7 +19,7 @@ SRC_URI="
 LICENSE="GPL-2-with-classpath-exception"
 KEYWORDS="~amd64"
 
-IUSE="alsa cups examples headless-awt selinux source webstart"
+IUSE="alsa cups examples headless-awt selinux source"
 
 RDEPEND="
 	media-libs/fontconfig:1.0
@@ -38,35 +38,47 @@ RDEPEND="
 		x11-libs/libXtst
 	)"
 
-PDEPEND="webstart? ( >=dev-java/icedtea-web-1.6.1:0 )"
-
 RESTRICT="preserve-libs strip"
 QA_PREBUILT="*"
 
 S="${WORKDIR}/zulu${MY_ZULU_PV}-${MY_ZULU_ARCH}"
 
+src_unpack() {
+	default
+	# 753575
+	if use arm; then
+		mv -v "${S}"* "${S}" || die
+	elif [[ ${A} == *_mac_* ]] ; then
+		mv -v "${S}/Contents/Home/"* "${S}" || die
+		rm -Rf "${S}/Contents"  # drop macOS executable
+	fi
+}
+
 src_install() {
 	local dest="/opt/${P}"
 	local ddest="${ED%/}/${dest#/}"
 
-	rm ASSEMBLY_EXCEPTION LICENSE THIRD_PARTY_README CLASSPATH_EXCEPTION_NOTE DISCLAIMER Welcome.html release readme.txt  || die
-	rm jre/ASSEMBLY_EXCEPTION jre/LICENSE jre/THIRD_PARTY_README || die
+	rm ASSEMBLY_EXCEPTION LICENSE THIRD_PARTY_README || die
 
-	# this does not exist on arm64 hence -f
-	rm -fv jre/lib/*/libfreetype.so* || die
+	# on macOS if they would exist they would be called .dylib, but most
+	# importantly, there are no different providers, so everything
+	# that's shipped works.
+	if [[ ${A} != *_mac_* ]] ; then
+		# this does not exist on arm64 hence -f
+		rm -fv jre/lib/*/libfreetype.so* || die
 
-	if ! use alsa ; then
-		rm -v jre/lib/*/libjsoundalsa.so* || die
-	fi
+		if ! use alsa ; then
+			rm -v jre/lib/*/libjsoundalsa.so* || die
+		fi
 
-	if ! use examples ; then
-		rm -vr demo || die
-		rm -vr sample || die
-	fi
+		if ! use examples ; then
+			rm -vr sample || die
+		fi
 
-	if use headless-awt ; then
-		rm -fvr {,jre/}lib/*/lib*{[jx]awt,splashscreen}* \
-			{,jre/}bin/policytool bin/appletviewer || die
+		if use headless-awt ; then
+			rm -fvr {,jre/}lib/*/lib*{[jx]awt,splashscreen}* \
+				{,jre/}bin/policytool bin/appletviewer || die
+		fi
 	fi
 
 	if ! use source ; then
